refactor components
Added components folder

Added file for each component

Added files to components folder
