+++ .vs/VSWorkspaceState.json
@@ -4,6 +4,6 @@
     "\\public",
     "\\src"
   ],
-  "SelectedNode": "\\src\\App.css",
+  "SelectedNode": "\\src\\data.json",
   "PreviewInSolutionExplorer": false
 }
