+++ src/components/TweetQuote.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+import '../App.css';
+
+
+const TweetQuote = (props) => {
+  return (
+    <a href="https://www.twitter.com/intent/tweet" rel="noopener noreferrer" target="_blank" id="tweet-quote" className="twitter fab fa-twitter-square fa-2x"></a>
+  );
+}
+
+export default TweetQuote;
